+++ common/src/main/java/io/telicent/smart/cache/storage/labels/CachingLabelsStore.java
@@ -5,7 +5,6 @@ package io.telicent.smart.cache.storage.labels;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.MapUtils;
 
 import java.util.*;