+++ hibernate/src/main/java/io/telicent/smart/cache/storage/hibernate/configuration/h2/H2Configuration.java
@@ -6,6 +6,7 @@ package io.telicent.smart.cache.storage.hibernate.configuration.h2;
 import io.telicent.smart.cache.storage.hibernate.configuration.DatabaseConfiguration;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 
 import java.io.File;
 import java.util.Properties;
@@ -21,8 +22,12 @@ public class H2Configuration {
      *
      * @param configuration Database configuration
      * @return Connection properties
+     * @throws IllegalArgumentException Thrown if the configuration is {@code null}, or invalid for use with an H2
+     *                                  in-memory backend
      */
     public static Properties prepareInMemoryConnectionProperties(DatabaseConfiguration configuration) {
+        validateConfiguration(configuration);
+
         Properties properties = new Properties();
         properties.put(JAKARTA_PERSISTENCE_JDBC_URL,
                        "jdbc:h2:mem:" + configuration.getDatabase() + ";DB_CLOSE_DELAY=-1");
@@ -33,6 +38,25 @@ public class H2Configuration {
     }
 
     /**
+     * Validates that the provided database configuration is valid for forming H2 database connections
+     *
+     * @param configuration Database configuration
+     * @throws IllegalArgumentException If the configuration is invalid in any way
+     */
+    private static void validateConfiguration(DatabaseConfiguration configuration) {
+        if (configuration == null) {
+            throw new IllegalArgumentException("Insufficient configuration provided for H2 database connections");
+        }
+        if (StringUtils.isNotBlank(configuration.getJdbcUrl())) {
+            throw new IllegalArgumentException(
+                    "Explicit JDBC URL configuration not permitted for H2 database connections");
+        }
+        if (StringUtils.isBlank(configuration.getDatabase())) {
+            throw new IllegalArgumentException("Must supply a database name for H2 database connections");
+        }
+    }
+
+    /**
      * Prepares database connection properties for an H2 file database
      *
      * @param configuration Database configuration
@@ -40,15 +64,19 @@ public class H2Configuration {
      *                      name provided in the {@link DatabaseConfiguration#getDatabase()} method relative to the base
      *                      directory.  If the base directory is {@code null} then it is ignored.
      * @return Connection properties
+     * @throws IllegalArgumentException Thrown if the configuration is {@code null}, or invalid for use with an H2 file
+     *                                  backend
      */
     public static Properties prepareFileConnectionProperties(DatabaseConfiguration configuration, File dbBaseDir) {
+        validateConfiguration(configuration);
+
         Properties properties = new Properties();
         properties.put(JAKARTA_PERSISTENCE_JDBC_URL,
                        "jdbc:h2:file:" + resolveDatabaseDirectory(configuration, dbBaseDir).getAbsolutePath());
         properties.put(JAKARTA_PERSISTENCE_JDBC_USER, "sa");
         properties.put(JAKARTA_PERSISTENCE_JDBC_PASSWORD, "");
         properties.put(HIBERNATE_DIALECT, "org.hibernate.dialect.H2Dialect");
-                return properties;
+        return properties;
     }
 
     /**