+++ hibernate/src/test/java/io/telicent/smart/cache/storage/hibernate/configuration/h2/TestH2Configuration.java
@@ -11,7 +11,42 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
+import static io.telicent.smart.cache.storage.hibernate.configuration.TestDatabaseConfiguration.FAKE_JDBC_URL;
+
 public class TestH2Configuration {
+
+    @Test(expectedExceptions = IllegalArgumentException.class)
+    public void givenNullConfiguration_whenPreparing_thenIllegalArgument() {
+        // Given, When and Then
+        H2Configuration.prepareInMemoryConnectionProperties(null);
+    }
+
+    @Test(expectedExceptions = IllegalArgumentException.class, expectedExceptionsMessageRegExp = "Explicit JDBC URL.*")
+    public void givenJdbcUrl_whenPreparingInMemoryConnectionProperties_thenIllegalArgument() {
+        // Given
+        DatabaseConfiguration configuration = DatabaseConfiguration.builder().jdbcUrl(FAKE_JDBC_URL).build();
+
+        // When and Then
+        H2Configuration.prepareInMemoryConnectionProperties(configuration);
+    }
+
+    @Test(expectedExceptions = IllegalArgumentException.class, expectedExceptionsMessageRegExp = "Explicit JDBC URL.*")
+    public void givenJdbcUrl_whenPreparingFileConnectionProperties_thenIllegalArgument() {
+        // Given
+        DatabaseConfiguration configuration = DatabaseConfiguration.builder().jdbcUrl(FAKE_JDBC_URL).build();
+
+        // When and Then
+        H2Configuration.prepareFileConnectionProperties(configuration, new File("."));
+    }
+
+    @Test(expectedExceptions = IllegalArgumentException.class, expectedExceptionsMessageRegExp = "Must supply a database name.*")
+    public void givenNoDatabaseName_whenPreparingConnectionProperties_thenIllegalArgument() {
+        // Given
+        DatabaseConfiguration configuration = DatabaseConfiguration.builder().build();
+
+        // When and Then
+        H2Configuration.prepareInMemoryConnectionProperties(configuration);
+    }
 
     @Test
     public void givenNoBaseDir_whenResolvingDbDir_thenUsedAsIs() {