+++ label-stores/label-store-benchmarks/src/main/java/io/telicent/smart/cache/storage/labels/benchmarks/VeryLargeLabelBenchmark.java
@@ -9,8 +9,6 @@ import org.apache.commons.lang3.RandomUtils;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.infra.Blackhole;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 /**